--- test/layers/presentation/chat/chat_viewmodel_test.mocks.dart
@@ -1,121 +1,0 @@
-// Mocks generated by Mockito 5.4.4 from annotations
-// in mobile1_flutter_coding_test/test/layers/presentation/chat/chat_viewmodel_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:mobile1_flutter_coding_test/layers/domain/authorization.dart'
-    as _i2;
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:mockito/src/dummies.dart' as _i3;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [Authorization].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockAuthorization extends _i1.Mock implements _i2.Authorization {
-  MockAuthorization() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  String get userId => (super.noSuchMethod(
-        Invocation.getter(#userId),
-        returnValue: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#userId),
-        ),
-      ) as String);
-
-  @override
-  set userId(String? _userId) => super.noSuchMethod(
-        Invocation.setter(
-          #userId,
-          _userId,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  String get name => (super.noSuchMethod(
-        Invocation.getter(#name),
-        returnValue: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#name),
-        ),
-      ) as String);
-
-  @override
-  set name(String? _name) => super.noSuchMethod(
-        Invocation.setter(
-          #name,
-          _name,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  String get email => (super.noSuchMethod(
-        Invocation.getter(#email),
-        returnValue: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#email),
-        ),
-      ) as String);
-
-  @override
-  set email(String? _email) => super.noSuchMethod(
-        Invocation.setter(
-          #email,
-          _email,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  String get profileImageURL => (super.noSuchMethod(
-        Invocation.getter(#profileImageURL),
-        returnValue: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#profileImageURL),
-        ),
-      ) as String);
-
-  @override
-  set profileImageURL(String? _profileImageURL) => super.noSuchMethod(
-        Invocation.setter(
-          #profileImageURL,
-          _profileImageURL,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  String get role => (super.noSuchMethod(
-        Invocation.getter(#role),
-        returnValue: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#role),
-        ),
-      ) as String);
-
-  @override
-  set role(String? _role) => super.noSuchMethod(
-        Invocation.setter(
-          #role,
-          _role,
-        ),
-        returnValueForMissingStub: null,
-      );
-}
